Rename servicedeploymanager files
This is to align ServiceDeployManager with the naming conventions
