+++ src/Orbit/main.cpp
@@ -46,9 +46,9 @@
 #include "Path.h"
 #include "SessionSetup/Connections.h"
 #include "SessionSetup/DeploymentConfigurations.h"
+#include "SessionSetup/ServiceDeployManager.h"
 #include "SessionSetup/SessionSetupDialog.h"
 #include "SessionSetup/TargetConfiguration.h"
-#include "SessionSetup/servicedeploymanager.h"
 #include "SourcePathsMapping/MappingManager.h"
 #include "Style/Style.h"
 #include "opengldetect.h"
