+++ src/OrbitQt/orbitmainwindow.cpp
@@ -92,9 +92,9 @@
 #include "QtUtils/MainThreadExecutorImpl.h"
 #include "SamplingReport.h"
 #include "SessionSetup/Connections.h"
+#include "SessionSetup/ServiceDeployManager.h"
 #include "SessionSetup/TargetConfiguration.h"
 #include "SessionSetup/TargetLabel.h"
-#include "SessionSetup/servicedeploymanager.h"
 #include "SourcePathsMapping/Mapping.h"
 #include "SourcePathsMapping/MappingManager.h"
 #include "SourcePathsMappingUI/AskUserForFile.h"
