+++ src/OrbitQt/orbitmainwindow.h
@@ -39,9 +39,9 @@
 #include "MainThreadExecutor.h"
 #include "MetricsUploader/MetricsUploader.h"
 #include "OrbitBase/CrashHandler.h"
+#include "SessionSetup/ServiceDeployManager.h"
 #include "SessionSetup/TargetConfiguration.h"
 #include "SessionSetup/TargetLabel.h"
-#include "SessionSetup/servicedeploymanager.h"
 #include "StatusListener.h"
 #include "capture_data.pb.h"
 #include "orbitglwidget.h"
