+++ src/SessionSetup/CMakeLists.txt
@@ -18,7 +18,7 @@ target_sources(
           include/SessionSetup/Error.h
           include/SessionSetup/OverlayWidget.h
           include/SessionSetup/ProcessItemModel.h
-          include/SessionSetup/servicedeploymanager.h
+          include/SessionSetup/ServiceDeployManager.h
           include/SessionSetup/SessionSetupDialog.h
           include/SessionSetup/TargetConfiguration.h
           include/SessionSetup/TargetLabel.h)
@@ -34,7 +34,7 @@ target_sources(
           OverlayWidget.cpp
           OverlayWidget.ui
           ProcessItemModel.cpp
-          servicedeploymanager.cpp
+          ServiceDeployManager.cpp
           SessionSetupDialog.cpp
           SessionSetupDialog.ui
           TargetLabel.cpp
